requirements.txt
apache-beam[gcp]
google-cloud-pubsub
google-cloud-bigtable
pandas==2.2.2